fix: update SQL query and improve error handling in booking API
- Modified the SQL query in the AppointmentsPage to cast default_fee to an integer for consistency.
- Enhanced error handling in the booking API to ensure proper formatting of unique constraint violation messages.
- Removed unused icons from the Sidebar component to streamline the layout.

These changes improve data integrity and user experience across the application.

+++ apps/web/src/app/(dashboard)/appointments/page.tsx
@@ -76,7 +76,7 @@ export default async function AppointmentsPage({
     }))
 
     doctors = await db.$queryRawUnsafe<Doctor[]>(
-      `SELECT id, name, speciality, default_fee FROM "${schemaName}".doctors ORDER BY name ASC`
+      `SELECT id, name, speciality, default_fee::int FROM "${schemaName}".doctors ORDER BY name ASC`
     )
   } catch {
     // Tenant schema may not be provisioned yet
